+++ extism-dev/exec.go
@@ -1,27 +1,24 @@
 package main
 
 import (
-	"context"
+	"os"
+	"os/exec"
 	"regexp"
 	"strings"
 
 	"github.com/extism/cli"
-	"github.com/extism/go-sdk"
 	"github.com/spf13/cobra"
 )
 
-type devCallArgs struct {
+type devExecArgs struct {
 	devArgs
 	category string
 	repo     string
+	shell    string
 	parallel int
-	timeout  int
 }
 
-func runDevCall(cmd *cobra.Command, args *devCallArgs) error {
-	plugin := args.args[0]
-	function := args.args[1]
-
+func runDevExec(cmd *cobra.Command, args *devExecArgs) error {
 	data, err := args.loadDataFile()
 	if err != nil {
 		return err
@@ -30,13 +27,6 @@ func runDevCall(cmd *cobra.Command, args *devCallArgs) error {
 	rx := &regexp.Regexp{}
 	if args.repo != "" {
 		rx = regexp.MustCompile(args.repo)
-	}
-
-	wasm := []extism.Wasm{}
-	if strings.HasPrefix(plugin, "http://") || strings.HasPrefix(plugin, "https://") {
-		wasm = append(wasm, extism.WasmUrl{Url: plugin})
-	} else {
-		wasm = append(wasm, extism.WasmFile{Path: plugin})
 	}
 
 	pool := NewPool(args.parallel)
@@ -52,32 +42,27 @@ func runDevCall(cmd *cobra.Command, args *devCallArgs) error {
 				}
 			}
 			p := repo.path()
-			cli.Log("Running plugin", plugin, "in", p)
+			cli.Log("Executing", args.args[0], "in", p, "using", args.shell)
 			if args.parallel <= 1 {
 				if i > 0 {
 					cli.Print()
 				}
 				cli.Print(p)
 			}
-			ctx := context.Background()
-			config := extism.PluginConfig{}
-			manifest := extism.Manifest{
-				Wasm:         wasm,
-				AllowedPaths: map[string]string{p: "/"},
+			a := strings.Join(args.args, " ")
+			cmd := exec.Command(args.shell, "-c", a)
+			cmd.Dir = p
+			cmd.Stdout = os.Stdout
+			cmd.Stderr = os.Stderr
+			cmd.Env = os.Environ()
+			cmd.Env = append(cmd.Env, "EXTISM_DEV_ROOT="+Root)
+			cmd.Env = append(cmd.Env, "EXTISM_DEV_RUNTIME="+args.Path("extism", "extism"))
+			cmd.Env = append(cmd.Env, "EXTISM_DEV_REPO_URL="+repo.Url)
+			cmd.Env = append(cmd.Env, "EXTISM_DEV_REPO_CATEGORY"+repo.Category)
+			cmd.Env = append(cmd.Env, "PATH="+os.Getenv("PATH")+":"+args.Path(".bin"))
+			if err := cmd.Run(); err != nil {
+				cli.Print("Error: command failed in", p)
 			}
-			if args.timeout != 0 {
-				manifest.Timeout = uint64(args.timeout)
-			}
-			plug, err := extism.NewPlugin(ctx, manifest, config, []extism.HostFunction{})
-			if err != nil {
-				cli.Print("Unable to create plugin:", err)
-				return
-			}
-			_, output, err := plug.Call(function, []byte(repo.Url))
-			if err != nil {
-				cli.Print("Plugin call failed in", p+":", err)
-			}
-			cli.Print(string(output))
 		}, r)
 	}
 	pool.Wait()
